* Fixed(?): Compilation issue in the Win32 code. The WinGr driver wrongly declared setCharacter. Bug tracker: SF #1607447.

+++ tvision/include/tv/wingr/screen.h
@@ -218,7 +218,7 @@ protected:
   static void  Resume();
   static void  Suspend();
   static void  setCharacter( unsigned offset
-                           , ushort value );
+                           , uint32 value );
 
   static void   setCharacters( unsigned dst
                              , ushort  *src
